Improve type-hints in .pyi stub
Summary:

Type-hints were too restrictive on APIs
which take a binding as a parameter. They didn't
account for Hashable values which can be used
instead of class types, as described in the docs.

Fixed some return values on the configure*()
APIs because they actually return the Injector.

Improved type-hint for the binding config Callable
on the configure*() APIs.

Made some type aliases to make it a little
easier to read.